provide wl_buffer abstraction
This patch moves the GEM abstraction into a separate file, which is used
by all the clients, instead of being in cairo-util.c for some clients
and in the client itself in other cases.

The presence of a global file descriptor variable, and the need to
open it separately from the creation of the wl_display, are fixed in a
subsequent patch.