add wl_backend, move opening of gem backend to wayland-client
This patch continues the refactoring by providing a separate wl_backend
object to be used by clients (now) and by compositors (later).
The wl_backend object is owned by the display; as a result of this
the wl_buffer API is changed by this patch to be methods of wl_display.

In subsequent patches, the server will be able to advertise which backend
it is using, so that the clients and compositors can agree and use the
same protocol (imagine the client using shm and the compositors using
GEM?...).