add backend to server-side wl_display
This is the next step in creating a common interface for compositors
and clients.  The server will now know the protocol on which the two
players have to agree, in terms of a driver and device (same as EGL
so that wl_display_get_egl_display works).

With this patch the EGL compositor is already using the information;
the other are using the device only (instead of hardcoding
/dev/dri/card0) and expecting the driver to be Intel.  The clients
already use the advertised device and driver, with no hard-coding
whatsoever.  The protocol used for advertising is simplistic, and will
be replaced by a proper capability object in the next patches.

On top of this, the patch adds advertisement of objects at connection
time.  A new server-exported object, wl_backend_advertisement,A is
used it to publish the chosen dbackend and arguments.  The client
accesses the information with a new API wl_display_get_interface
which, for now, is invoked only by wl_display_create.  The client
could also use the API to query for the availability of input devices,
as will be done in the next patch.

Besides the advertisement mechanism and the new object, there is a bug
fix in that packages of length exactly 8 were rejected; these correspond
to methods with no arguments.