GLX compositor: big endian support
The remaining patches tweak glx-compositor so that it would work in my
admittedly borderline testing.  First of all, support big-endian clients
because Cairo stores things as RGBA there instead of ARGB (or is it the
other way round?...).

This may be done more cleanly after the package is autoconfiscated.
I didn't touch the EGL compositor.