initial shm implementation
This is an initial implementation of the shared memory backend, using
Doug Lea's malloc to manage memory allocation.

It is quite incomplete in that it requires clients to free their own
memory (fixing this would probably need a help from the server) and
it is not thread-safe (since it's across multiple process, the easiest
thing to do is to use a lock-free queue, but I don't have enough time
to do that now...).

It does not support EGL of course, so it only works with glx-compositor
and does not work with the window.c client.  As a toy, and as a way to
enhance portability beyond Intel chipsets, it is great though.