move demarshaling function to connection.c
This starts adding marshal/demarshal tools to connection.c so that
the server and libwayland.so both can use them.

It disables WL_ARGUMENT_INTERFACE because objects in libwayland.so
unfortunately do not know their interface name.  This could be
fixed easily, but currently requires care to keep the same binary
layout of wl_object and wl_proxy.  Since there is 'o' I left this
aside.